+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/entity/create/StaffEntity.kt
@@ -5,6 +5,7 @@ import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.Id
 import jakarta.persistence.Table
+import java.math.BigInteger
 import java.util.UUID
 
 @Entity
@@ -16,7 +17,7 @@ data class StaffEntity(
   var id: UUID? = null,
 
   @Column(name = "staff_id")
-  var staffId: Int? = null,
+  var staffId: BigInteger? = null,
 
   @Column(name = "first_name", nullable = false)
   var firstName: String,
